Delete -aux | grep -e "mongo" | grep -v "grep"
I somehow created this file a while back accidentally, for some reason after trying to execute this command it actually created a file name though I don't remember putting "touch," before it, not really sure what happened. Deleting it now because it's not serving any purpose and it appears that windows prevents cloning a directory containing files with a special character in its name.