Update Volume-47-P5.xml
I have fixed the XML by moving TC's Will so that it is appropriately situated as back matter and tagged not as a letter but as an appendix. This  should return the letters themselves to their proper place. Need to check the appendix coding to make sure there is no conflict and that the sourcenote / footnotes display properly -- we can change that to another kind of heard if it will not work.